--- a/src/structs/ansi_style_properties.h
+++ b/src/structs/ansi_style_properties.h
@@ -5,19 +5,26 @@
 
 struct ansi_style_properties
 {
-    bool bold : 1;
-    bool faint : 1;
-    bool italic : 1;
-    bool underline : 1;
-    bool slow_blink : 1;
-    bool fast_blink : 1;
-    bool reverse : 1;
-    bool crossout : 1;
-    bool fraktur : 1;
-    bool double_underline : 1;
-    bool frame : 1;
-    bool circle : 1;
-    bool overline : 1;
+    union
+    {
+        struct
+        {
+            bool bold : 1;
+            bool faint : 1;
+            bool italic : 1;
+            bool underline : 1;
+            bool slow_blink : 1;
+            bool fast_blink : 1;
+            bool reverse : 1;
+            bool crossout : 1;
+            bool fraktur : 1;
+            bool double_underline : 1;
+            bool frame : 1;
+            bool circle : 1;
+            bool overline : 1;
+        };
+        unsigned short any;
+    };
 };
 
 #endif // ANSI_STYLE_PROPERTIES_H