A couple reasonable things to speed things up a fair bit.

Given the usual test file:

```console
$ wc -c foo100.txt ; sak stripansi < foo100.txt | wc -c ; ./ansi2html < foo100.txt | wc -c ; ./ansi2html --use-classes < foo100.txt | wc -c ;
517379100 foo100.txt
167182500
1133143500
860183700
```

... there's a decent speed-up between master and this branch:

```console
$ hyperfine --time-unit millisecond --warmup 5 './ansi2html-master < foo100.txt' './ansi2html-branch < foo100.txt'
Benchmark 1: ./ansi2html-master < foo100.txt
  Time (mean ± σ):     1031.2 ms ±  16.3 ms    [User: 985.8 ms, System: 41.9 ms]
  Range (min … max):   1015.2 ms … 1063.2 ms    10 runs

Benchmark 2: ./ansi2html-branch < foo100.txt
  Time (mean ± σ):     982.6 ms ±  10.1 ms    [User: 943.6 ms, System: 34.9 ms]
  Range (min … max):   971.5 ms … 1002.8 ms    10 runs

Summary
  './ansi2html-branch < foo100.txt' ran
    1.05 ± 0.02 times faster than './ansi2html-master < foo100.txt'
```

... which can also be seen via pv's MB/s:

```console
$ ./ansi2html-master < foo100.txt | pv >/dev/null
1.06GiB 0:00:01 [ 773MiB/s] [           <=>   ]
$ ./ansi2html-branch < foo100.txt | pv >/dev/null
1.06GiB 0:00:01 [ 822MiB/s] [           <=>   ]
```